+++ main.go
@@ -2,57 +2,12 @@ package main
 
 import (
 	"aifa-uni-bonn/home-reverse-proxy/doproxy"
-	"errors"
-	"fmt"
 	"log"
 	"net/http"
-	"net/http/httputil"
-	"net/url"
 )
 
-// NewProxy takes target host and creates a reverse proxy
-func NewProxy(targetHost string) (*httputil.ReverseProxy, error) {
-	url, err := url.Parse(targetHost)
-	if err != nil {
-		return nil, err
-	}
-
-	proxy := httputil.NewSingleHostReverseProxy(url)
-
-	originalDirector := proxy.Director
-	proxy.Director = func(req *http.Request) {
-		originalDirector(req)
-		modifyRequest(req)
-	}
-
-	//proxy.ModifyResponse = modifyResponse()
-	proxy.ErrorHandler = errorHandler()
-	return proxy, nil
-}
-
-func modifyRequest(req *http.Request) {
-	req.Header.Set("X-Proxy", "Simple-Reverse-Proxy")
-}
-
-func errorHandler() func(http.ResponseWriter, *http.Request, error) {
-	return func(w http.ResponseWriter, req *http.Request, err error) {
-		fmt.Printf("Got error while modifying response: %v \n", err)
-		return
-	}
-}
-
-func modifyResponse() func(*http.Response) error {
-	return func(resp *http.Response) error {
-		return errors.New("response body is invalid")
-	}
-}
-
-// ProxyRequestHandler handles the http request using proxy
-func ProxyRequestHandler(proxy *httputil.ReverseProxy) func(http.ResponseWriter, *http.Request) {
-	return doproxy.Handle_proxy_request
-}
-
 func main() {
+	doproxy.Init_doproxy()
 	log.SetFlags(log.Lshortfile | log.LstdFlags)
 	// initialize a reverse proxy and pass the actual backend server url here
 	//proxy, err := NewProxy("http://web-www2019.astro.uni-bonn.de")
